Handle code review feedback.

+++ pkg/controller/che/che_controller.go
@@ -383,17 +383,31 @@ func (r *ReconcileChe) Reconcile(request reconcile.Request) (reconcile.Result, e
 		}
 	}
 
-	if isOpenShift4 && util.IsDeleteOAuthInitialUser(instance) {
+	if isOpenShift4 && instance.Spec.Auth.InitialOpenShiftOAuthUser != nil && util.IsDeleteOAuthInitialUser(instance) {
 		if err := r.userHandler.DeleteOAuthInitialUser(deployContext); err != nil {
 			logrus.Errorf("Unable to delete initial OpenShift OAuth user from a cluster. Cause: %s", err.Error())
 			instance.Spec.Auth.InitialOpenShiftOAuthUser = nil
 			err := r.UpdateCheCRSpec(instance, "InitialOpenShiftOAuthUser", "nil")
-			return reconcile.Result{RequeueAfter: time.Second * 1}, err
+			return reconcile.Result{}, err
 		}
+
+		instance.Spec.Auth.OpenShiftoAuth = nil
+		if err := r.UpdateCheCRSpec(instance, "OpenShiftoAuth", "nil"); err != nil {
+			return reconcile.Result{}, err
+		}
+
+		instance.Spec.Auth.InitialOpenShiftOAuthUser = nil
+		err := r.UpdateCheCRSpec(instance, "InitialOpenShiftOAuthUser", "nil")
+		if err != nil {
+			return reconcile.Result{}, err
+		}
+
 		instance.Status.OpenShiftOAuthUserCredentialsSecret = ""
 		if err := r.UpdateCheCRStatus(instance, "openShiftOAuthUserCredentialsSecret", openShiftOAuthUserCredentialsSecret); err != nil {
 			return reconcile.Result{}, err
 		}
+
+		return reconcile.Result{}, nil
 	}
 
 	if isOpenShift && instance.Spec.Auth.OpenShiftoAuth == nil {
@@ -1134,7 +1148,8 @@ func (r *ReconcileChe) autoEnableOAuth(deployContext *deploy.DeployContext, requ
 			if len(openshitOAuth.Spec.IdentityProviders) > 0 {
 				oauth = true
 			} else if util.IsInitialOpenShiftOAuthUserEnabled(cr) {
-				if err := r.userHandler.CreateOAuthInitialUser(openshitOAuth, deployContext); err != nil {
+				provisioned, err := r.userHandler.SyncOAuthInitialUser(openshitOAuth, deployContext);
+				if err != nil {
 					message = warningNoIdentityProvidersMessage + " Operator tried to create initial OpenShift OAuth user for HTPasswd identity provider, but failed. Cause: " + err.Error()
 					logrus.Error(message)
 					logrus.Info("To enable OpenShift OAuth, please add identity provider first: " + howToAddIdentityProviderLinkOS4)
@@ -1153,6 +1168,9 @@ func (r *ReconcileChe) autoEnableOAuth(deployContext *deploy.DeployContext, requ
 							return reconcile.Result{}, err
 						}
 					}
+				}
+				if !provisioned {
+					return reconcile.Result{}, err
 				}
 			}
 		}
